+++ xcode/MAME4iOS/TVOptionsController.m
@@ -9,9 +9,9 @@
 #import "TVOptionsController.h"
 #import "ListOptionController.h"
 #import "TVInputOptionsController.h"
-#import "SystemImage.h"
 #import "CloudSync.h"
 #import "Alert.h"
+#import "MAME4iOS-Swift.h"
 
 @implementation TVOptionsController